+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+- A new option `--follow` was added to specify whether `todos` should follow
+  symlinks when traversing directories.
+  ([#1752](https://github.com/ianlewis/todos/issues/1752)).
+
 ## [`0.13.0`] - 2025-05-18
 
 ### Added in `0.13.0`